We made a bad call, and ended up integrating a testing branch into ou… (#52)
…r primary frontend branch. As such we have decided to do a manual
rebase, copying the up-to-date frontend code, into a new dev-based
branch. This will result in a horrible commit tree, and we do
appologise. The following is all the code from the first two sprints,
compiled into a single commit.

# Pull Request (PR) Template

## Description
Please include a concise summary of the change and the problem it
solves. Detail the motivation behind the changes, link to any related
issues, and provide context for the reviewers.
To check a box simply put an X.

## Type of Change
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation Update
- [x] Other (please describe)

As mentioned above, this pull-request entails two whole sprints of work
from frontend.

## Checklist
Before submitting your PR, please ensure the following:
- [x] I have performed a self-review of my own code.
- [x] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have made corresponding changes to the documentation (if
appropriate).
- [x] My changes generate no new warnings.
- [ ] I have added tests that prove my fix is effective or that my
feature works.
- [x] All new and existing tests passes.

## Screenshots (if appropriate)

## Additional Notes
Include any additional information that you believe is important for the
reviewers to know.