frontend/TournamentService.JS fix (#107)
# Pull Request (PR) Template

## Description
Please include a concise summary of the change and the problem it
solves. Detail the motivation behind the changes, link to any related
issues, and provide context for the reviewers.
To check a box simply put an X.

## Type of Change
- [ ] New feature (non-breaking change which adds functionality)
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation Update
- [ ] Other (please describe)

## Checklist
Before submitting your PR, please ensure the following:
- [x] I have performed a self-review of my own code.
- [ ] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have made corresponding changes to the documentation (if
appropriate).
- [x] My changes generate no new warnings.
- [ ] I have added tests that prove my fix is effective or that my
feature works.
- [ ] All new and existing tests passes.

## Screenshots (if appropriate)

## Additional Notes
Include any additional information that you believe is important for the
reviewers to know.