Iot/feature/85 send timestamp (#99)
# Pull Request (PR) Template
This is not an actual PR. It serves as a draft for now. The feature is
not yet implemented, but the groundwork has been laid. I expect to
finish it tomorrow (sometime) or in the weekend.

## Description
Please include a concise summary of the change and the problem it
solves. Detail the motivation behind the changes, link to any related
issues, and provide context for the reviewers.
To check a box simply put an X.

## Type of Change
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation Update
- [ ] Other (please describe)

## Checklist
Before submitting your PR, please ensure the following:
- [ ] I have performed a self-review of my own code.
- [ ] I have commented my code, particularly in hard-to-understand
areas.
- [ ] I have made corresponding changes to the documentation (if
appropriate).
- [ ] My changes generate no new warnings.
- [ ] I have added tests that prove my fix is effective or that my
feature works.
- [ ] All new and existing tests passes.

## Screenshots (if appropriate)

## Additional Notes
Include any additional information that you believe is important for the
reviewers to know.

---------

Co-authored-by: Rune Lauridsen <rune.lauridsen.2@gmail.com>